+++ aldrin/src/low_level.rs
@@ -4,7 +4,6 @@
 
 mod call;
 mod event;
-mod event_listener;
 mod promise;
 mod proxy;
 mod reply;
@@ -12,12 +11,10 @@ mod service;
 #[cfg(test)]
 mod test;
 
-pub(crate) use event_listener::{EventListenerId, EventListenerRequest};
 pub(crate) use service::RawCall;
 
 pub use call::Call;
 pub use event::Event;
-pub use event_listener::EventListener;
 pub use promise::Promise;
 pub use proxy::Proxy;
 pub use reply::Reply;