+++ aldrin/src/low_level/proxy.rs
@@ -1,5 +1,5 @@
+use super::Event;
 use super::Reply;
-use super::{Event, EventListener};
 #[cfg(feature = "introspection")]
 use crate::core::introspection::Introspection;
 use crate::core::{Serialize, ServiceId, ServiceInfo, TypeId};
@@ -18,20 +18,13 @@ pub struct Proxy {
     client: Handle,
     id: ServiceId,
     info: ServiceInfo,
-    events: Option<EventListener>,
 }
 
 impl Proxy {
     /// Creates a new proxy to a service.
     pub async fn new(client: Handle, id: ServiceId) -> Result<Self, Error> {
         let info = client.query_service_info(id).await?;
-
-        Ok(Self {
-            client,
-            id,
-            info,
-            events: None,
-        })
+        Ok(Self { client, id, info })
     }
 
     /// Returns a handle to the proxy's client.
@@ -75,30 +68,21 @@ impl Proxy {
     ///
     /// This function returns `true`, if `event` was not already subscribed to. Otherwise `false` is
     /// returned.
-    pub async fn subscribe_event(&mut self, event: u32) -> Result<bool, Error> {
-        self.events
-            .get_or_insert_with(|| EventListener::new(self.client.clone()))
-            .subscribe(self.id, event)
-            .await
+    pub async fn subscribe_event(&mut self, _event: u32) -> Result<bool, Error> {
+        todo!()
     }
 
     /// Unsubscribe from an event.
     ///
     /// This function returns `true`, if `event` was subscribed to before the call to this function
     /// and is now unsubscribed from. Otherwise `false` is returned.
-    pub fn unsubscribe_event(&mut self, event: u32) -> Result<bool, Error> {
-        match self.events {
-            Some(ref mut listener) => listener.unsubscribe(self.id, event),
-            None => Ok(false),
-        }
+    pub fn unsubscribe_event(&mut self, _event: u32) -> Result<bool, Error> {
+        todo!()
     }
 
     /// Polls for the next event.
-    pub fn poll_next_event(&mut self, cx: &mut Context) -> Poll<Option<Event>> {
-        match self.events {
-            Some(ref mut listener) => listener.poll_next_event(cx),
-            None => Poll::Ready(None),
-        }
+    pub fn poll_next_event(&mut self, _cx: &mut Context) -> Poll<Option<Event>> {
+        todo!()
     }
 
     /// Returns the next event.
@@ -111,10 +95,7 @@ impl Proxy {
     /// When `events_finished` returns `true`, then [`next_event`](Self::next_event) is guaranteed
     /// to return `None`.
     pub fn events_finished(&self) -> bool {
-        match self.events {
-            Some(ref listener) => listener.is_finished(),
-            None => false,
-        }
+        todo!()
     }
 }